feat(engine): extract buildIssueRagQuery to gittensory-engine
buildIssueRagQuery is pure, string-only query composition (#2320) that
the miner analyze phase needs before any PR diff exists, but it lived in
src/review/issue-rag-wire.ts where the miner cannot import it. This
moves it to packages/gittensory-engine/src/issue-rag-query.ts and turns
issue-rag-wire.ts into a thin re-export shim, following the same layout
as the #2278 duplicate-winner and #2282 scoring extractions.

MIN_QUERY_CHARS moves with it: the constant is the floor both the query
builder and rag.ts's retrieval guard check, so it is defined once in the
engine module and re-exported from src/review/rag.ts — keeping a single
source of truth instead of two drifting copies. The Vectorize/D1-bound
retrieval backend (retrieveContext) intentionally stays in src, exactly
as the issue scopes it.

Exported from the package entrypoint, covered by an engine-side barrel
test mirroring the duplicate-winner extraction's, and the existing
test/unit/issue-rag-wire.test.ts passes unmodified through the shim.

Closes #4254
