+++ src/review/issue-rag-wire.ts
@@ -1,38 +1,10 @@
-// Issue-centric RAG query composition (#2320). The miner analyze phase has no PR diff yet, so it feeds retrieval
-// from the issue's title/body/labels while reusing the existing RAG engine unchanged.
-
-import { MIN_QUERY_CHARS } from "./rag";
-
-const MAX_ISSUE_BODY_CHARS = 4000;
-const MAX_ISSUE_LABELS = 20;
-
-export type IssueRagQueryInput = {
-  title: string;
-  body?: string | undefined;
-  labels?: string[] | undefined;
-};
-
-function cleanLabels(labels: string[] | undefined): string[] {
-  if (!labels) return [];
-  return labels
-    .map((label) => label.trim())
-    .filter(Boolean)
-    .slice(0, MAX_ISSUE_LABELS);
-}
-
-// Intentionally pre-built and currently unreached: no miner-side issue-analysis caller exists yet (#2320).
-export function buildIssueRagQuery(input: IssueRagQueryInput): { queryText: string } {
-  const sections: string[] = [];
-  const title = input.title.trim();
-  if (title) sections.push(title);
-
-  const body = (input.body ?? "").trim().slice(0, MAX_ISSUE_BODY_CHARS);
-  if (body) sections.push(body);
-
-  const labels = cleanLabels(input.labels);
-  if (labels.length > 0) sections.push(`Labels: ${labels.join(", ")}`);
-
-  const queryText = sections.join("\n\n").trim();
-  if (queryText.length < MIN_QUERY_CHARS) return { queryText: "" };
-  return { queryText };
-}
+/**
+ * Issue-centric RAG query composition (#2320), extracted to `@jsonbored/gittensory-engine` (#4254) so the
+ * miner analyze phase can compose the identical retrieval query from an issue's title/body/labels without
+ * importing the review stack. The retrieval backend itself (`retrieveContext` in `./rag`) is Vectorize/D1-bound
+ * and intentionally stays in `src` — this shim only re-exports the pure query builder.
+ *
+ * packages/gittensory-engine/src/issue-rag-query.ts (imported via relative source path, not the published
+ * module, matching the #2278/#2282 extraction shims) is the source of truth.
+ */
+export { buildIssueRagQuery, type IssueRagQueryInput } from "../../packages/gittensory-engine/src/issue-rag-query";
